requirements.txt
pygame
pymongo
sentence-transformers
numpy
python-dotenv
pillow
